Reconstruct lexer part 1, remove dbtype judgement from root lexer.
