+++ core/src/main/java/com/alibaba/druid/sql/dialect/hive/parser/HiveLexer.java
@@ -18,6 +18,7 @@ package com.alibaba.druid.sql.dialect.hive.parser;
 import com.alibaba.druid.DbType;
 import com.alibaba.druid.sql.parser.Keywords;
 import com.alibaba.druid.sql.parser.Lexer;
+import com.alibaba.druid.sql.parser.NotAllowCommentException;
 import com.alibaba.druid.sql.parser.ParserException;
 import com.alibaba.druid.sql.parser.SQLParserFeature;
 import com.alibaba.druid.sql.parser.Token;
@@ -25,6 +26,8 @@ import com.alibaba.druid.sql.parser.Token;
 import java.util.HashMap;
 import java.util.Map;
 
+import static com.alibaba.druid.sql.parser.CharTypes.isWhitespace;
+import static com.alibaba.druid.sql.parser.LayoutCharacters.EOI;
 import static com.alibaba.druid.sql.parser.Token.LITERAL_CHARS;
 
 public class HiveLexer extends Lexer {
@@ -80,7 +83,7 @@ public class HiveLexer extends Lexer {
     }
 
     @Override
-    protected final void scanString() {
+    protected void scanString() {
         {
             boolean hasSpecial = false;
             int startIndex = pos + 1;
@@ -234,10 +237,156 @@ public class HiveLexer extends Lexer {
         }
     }
 
-    public void scanComment() {
-        scanHiveComment();
+    protected boolean supportScanHiveCommentDoubleSpace() {
+        return false;
     }
 
+    public void scanComment() {
+        if (ch != '/' && ch != '-') {
+            throw new IllegalStateException();
+        }
+
+        Token lastToken = this.token;
+
+        mark = pos;
+        bufPos = 0;
+        scanChar();
+
+        if (ch == ' ') {
+            mark = pos;
+            bufPos = 0;
+            scanChar();
+
+            if (supportScanHiveCommentDoubleSpace() && ch == ' ') {
+                mark = pos;
+                bufPos = 0;
+                scanChar();
+            }
+        }
+
+        // /*+ */
+        if (ch == '*') {
+            scanChar();
+            bufPos++;
+
+            while (ch == ' ') {
+                scanChar();
+                bufPos++;
+            }
+
+            boolean isHint = false;
+            int startHintSp = bufPos + 1;
+            if (ch == '+') {
+                isHint = true;
+                scanChar();
+                bufPos++;
+            }
+
+            for (; ; ) {
+                if (ch == '*') {
+                    if (charAt(pos + 1) == '/') {
+                        bufPos += 2;
+                        scanChar();
+                        scanChar();
+                        break;
+                    } else if (isWhitespace(charAt(pos + 1))) {
+                        int i = 2;
+                        for (; i < 1024 * 1024; ++i) {
+                            if (!isWhitespace(charAt(pos + i))) {
+                                break;
+                            }
+                        }
+                        if (charAt(pos + i) == '/') {
+                            bufPos += 2;
+                            pos += (i + 1);
+                            ch = charAt(pos);
+                            break;
+                        }
+                    }
+                }
+
+                scanChar();
+                if (ch == EOI) {
+                    break;
+                }
+                bufPos++;
+            }
+
+            if (isHint) {
+                stringVal = subString(mark + startHintSp, (bufPos - startHintSp) - 1);
+                token = Token.HINT;
+            } else {
+                stringVal = subString(mark, bufPos + 1);
+                token = Token.MULTI_LINE_COMMENT;
+                commentCount++;
+                if (keepComments) {
+                    addComment(stringVal);
+                }
+            }
+
+            if (commentHandler != null && commentHandler.handle(lastToken, stringVal)) {
+                return;
+            }
+
+            if (token != Token.HINT && !isAllowComment()) {
+                throw new NotAllowCommentException();
+            }
+
+            return;
+        }
+
+        if (!isAllowComment()) {
+            throw new NotAllowCommentException();
+        }
+
+        if (ch == '/' || ch == '-') {
+            scanChar();
+            bufPos++;
+
+            for (; ; ) {
+                if (ch == '\r') {
+                    if (charAt(pos + 1) == '\n') {
+                        line++;
+                        bufPos += 2;
+                        scanChar();
+                        break;
+                    }
+                    bufPos++;
+                    break;
+                } else if (ch == EOI) {
+                    if (pos >= text.length()) {
+                        break;
+                    }
+                }
+
+                if (ch == '\n') {
+                    line++;
+                    scanChar();
+                    bufPos++;
+                    break;
+                }
+
+                scanChar();
+                bufPos++;
+            }
+
+            stringVal = subString(mark, ch != EOI ? bufPos : bufPos + 1);
+            token = Token.LINE_COMMENT;
+            commentCount++;
+            if (keepComments) {
+                addComment(stringVal);
+            }
+            endOfComment = isEOF();
+
+            if (commentHandler != null && commentHandler.handle(lastToken, stringVal)) {
+                return;
+            }
+
+            return;
+        }
+    }
+
+    @Override
     protected boolean supportNextTokenColon() {
         return true;
     }
