+++ core/src/main/java/com/alibaba/druid/sql/parser/Lexer.java
@@ -2333,155 +2333,6 @@ public class Lexer {
         }
     }
 
-    protected boolean supportScanHiveCommentDoubleSpace() {
-        return false;
-    }
-
-    protected final void scanHiveComment() {
-        if (ch != '/' && ch != '-') {
-            throw new IllegalStateException();
-        }
-
-        Token lastToken = this.token;
-
-        mark = pos;
-        bufPos = 0;
-        scanChar();
-
-        if (ch == ' ') {
-            mark = pos;
-            bufPos = 0;
-            scanChar();
-
-            if (supportScanHiveCommentDoubleSpace() && ch == ' ') {
-                mark = pos;
-                bufPos = 0;
-                scanChar();
-            }
-        }
-
-        // /*+ */
-        if (ch == '*') {
-            scanChar();
-            bufPos++;
-
-            while (ch == ' ') {
-                scanChar();
-                bufPos++;
-            }
-
-            boolean isHint = false;
-            int startHintSp = bufPos + 1;
-            if (ch == '+') {
-                isHint = true;
-                scanChar();
-                bufPos++;
-            }
-
-            for (; ; ) {
-                if (ch == '*') {
-                    if (charAt(pos + 1) == '/') {
-                        bufPos += 2;
-                        scanChar();
-                        scanChar();
-                        break;
-                    } else if (isWhitespace(charAt(pos + 1))) {
-                        int i = 2;
-                        for (; i < 1024 * 1024; ++i) {
-                            if (!isWhitespace(charAt(pos + i))) {
-                                break;
-                            }
-                        }
-                        if (charAt(pos + i) == '/') {
-                            bufPos += 2;
-                            pos += (i + 1);
-                            ch = charAt(pos);
-                            break;
-                        }
-                    }
-                }
-
-                scanChar();
-                if (ch == EOI) {
-                    break;
-                }
-                bufPos++;
-            }
-
-            if (isHint) {
-                stringVal = subString(mark + startHintSp, (bufPos - startHintSp) - 1);
-                token = Token.HINT;
-            } else {
-                stringVal = subString(mark, bufPos + 1);
-                token = Token.MULTI_LINE_COMMENT;
-                commentCount++;
-                if (keepComments) {
-                    addComment(stringVal);
-                }
-            }
-
-            if (commentHandler != null && commentHandler.handle(lastToken, stringVal)) {
-                return;
-            }
-
-            if (token != Token.HINT && !isAllowComment()) {
-                throw new NotAllowCommentException();
-            }
-
-            return;
-        }
-
-        if (!isAllowComment()) {
-            throw new NotAllowCommentException();
-        }
-
-        if (ch == '/' || ch == '-') {
-            scanChar();
-            bufPos++;
-
-            for (; ; ) {
-                if (ch == '\r') {
-                    if (charAt(pos + 1) == '\n') {
-                        line++;
-                        bufPos += 2;
-                        scanChar();
-                        break;
-                    }
-                    bufPos++;
-                    break;
-                } else if (ch == EOI) {
-                    if (pos >= text.length()) {
-                        break;
-                    }
-                }
-
-                if (ch == '\n') {
-                    line++;
-                    scanChar();
-                    bufPos++;
-                    break;
-                }
-
-                scanChar();
-                bufPos++;
-            }
-
-            stringVal = subString(mark, ch != EOI ? bufPos : bufPos + 1);
-            token = Token.LINE_COMMENT;
-            commentCount++;
-            if (keepComments) {
-                addComment(stringVal);
-            }
-            endOfComment = isEOF();
-
-            if (commentHandler != null && commentHandler.handle(lastToken, stringVal)) {
-                return;
-            }
-
-            return;
-        }
-    }
-
     public List<String> scanLineArgument() {
         List<String> args = new ArrayList<>();
         while (ch == ' ') {
